refactor: split adapter cli reference helpers
